// repository/frontend/src/app/houseplans/interface_Houseplans/data-houseplans-interface.ts
export interface DataHouseplans {
  // Common
  id_Houseplans: number;
  name_Houseplans: string;
  image_Houseplans: string;
  price_Houseplans: number;
  rating_Houseplans: number;
  sqfttotal_Houseplans: number;
  wide_Houseplans: string;
  deep_Houseplans: string;
  story_Houseplans: number;
  bed_Houseplans: number;
  bath_Houseplans: number;
  basement_Houseplans: boolean;
  garage_Houseplans: boolean;

  // Product details

  // Basket details
}